refactor utils to converter module
